requirements.txt
SQLAlchemy==2.0.49
altair==6.1.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==7.0.6
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
curl_cffi==0.15.0
flask==3.1.1
frozendict==2.4.7
greenlet==3.5.0
gunicorn==23.0.0
idna==3.13
jsonschema==4.26.0
markdown-it-py==4.0.0
multitasking==0.0.13
narwhals==2.20.0
numpy==2.4.4
packaging==26.2
pandas==3.0.2
pillow==12.2.0
plotly==6.7.0
protobuf==7.34.1
psycopg2-binary==2.9.12
pyarrow==24.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.33.1
rich==15.0.0
scipy==1.17.1
six==1.17.0
streamlit==1.57.0
tenacity==9.1.4
toml==0.10.2
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3
yfinance==1.3.0
